marine_mapping/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-gis==1.1
django-cors-headers==4.4.0
requests==2.32.3
